安小收全部接口

+++ src/main/java/com/gbq/axs/axsxcs/pojo/ResPageBean.java
@@ -17,7 +17,7 @@ import java.util.List;
 @AllArgsConstructor
 public class ResPageBean {
     //总条数
-    private Integer total;
+    private long total;
     //数据list
     private List<?> data;
 }
